requirements.txt
pytest==8.2.2
requests==2.32.3
selene==2.0.0rc9
allure-pytest==2.13.5
python-dotenv==1.0.1
